+++ scripts/check-version-consistency.sh
@@ -80,6 +80,7 @@ stale_versions=(
   "v0.5.2"
   "v0.5.3"
   "v0.5.4"
+  "v0.5.5"
 )
 
 entry_files=(